Use `packageSourceOverrides` (#3)
The original reason for not using this was to
work around a mistake on our end.  When we specified
`nix` via the `extra-libraries` clause of the `.cabal`
file instead of `pkgconfig-depends` that messed up the
file generated by `cabal2nix` which entailed various
work-arounds.  Those workarounds are no longer necessary,
so we can switch to `packageSourceOverrides`